Make project build on mac as well
Linked allegro_main and fixed main() signature to let allegro use
mangled main function. Otherwise it would not work on a Mac.
Also typedef std::u32string away as a basic_with with char32_t
when on Mac - excuse the poor with old compilers. Also, rand()
function needs <cstdlib> to be sourced for the compiler to find it.